add in get balance

+++ context.go
@@ -2,5 +2,5 @@ package solana
 
 // Context is extra contextual information for RPC responses
 type Context struct {
-	Slot int `json:"slot"`
+	Slot uint64 `json:"slot"`
 }
